requirements.txt
autopep8==2.0.0
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
python-decouple==3.6
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2